modules/astra-ingest/pyproject.toml
[tool.poetry]
name = "astra-ingest"
version = "0.1.0"
description = "ETL & Ingestion Engine for ASTRA Platform"
authors = ["Astra Engineering Team"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
# Base de Datos Relacional
sqlalchemy = "^2.0.23"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.0"
# Base de Datos Vectorial
qdrant-client = "^1.7.1"
# Procesamiento y Utilidades
pydantic = "^2.5.2"
python-dotenv = "^1.0.0"
lxml = "^5.2.0"
pydantic-settings = "^2.12.0"
# Nuevas dependencias para AssetLibrary (T03)
Pillow = "^10.1.0"
ImageHash = "^4.3.1"
grpcio = "^1.59.3"
grpcio-tools = "^1.59.3"
# NLP / Vectorización (T05)
sentence-transformers = "^2.2.2"
torch = ">=2.0.0"
scikit-learn = "^1.3.0"
python-docx = "^1.2.0"
spacy = "^3.8.11"
fastapi = "^0.129.0"
uvicorn = "^0.40.0"
tabulate = "^0.9.0"
psutil = "^5.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.11.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["src", "tests"]